fair_queue: Move throttling code to io_queue::stream class
Continuation of the previous patch. Now io_queue calls throttler
capacity grabbing via fair_queue. This patch relocates the grabbing
methods to io_queue::stream. For that to work some helper fields from
fair_queue need to be moved to the stream as well.
